+++ sdk/ml/azure-ai-ml/tests/workspace/unittests/test_workspace_operations_base.py
@@ -136,6 +136,7 @@ class TestWorkspaceOperation:
                     ),
                 ],
             )
+            ws.provision_network_now = True
             ws.system_datastores_auth_mode = "identity"
             ws.allow_roleassignment_on_rg = True
             return ws._to_rest_object()
@@ -162,6 +163,8 @@ class TestWorkspaceOperation:
         assert rules[2].service_tag == "sometag"
         assert rules[2].protocol == "*"
         assert rules[2].port_ranges == "1,2"
+
+        assert ws.provision_network_now == True
 
     def test_create_get_exception_swallow(
         self,